Refactoring of the tests folder for the trainers
