+++ ml-agents/mlagents/trainers/tests/tensorflow/test_bcmodule.py
@@ -2,7 +2,6 @@ import pytest
 import mlagents.trainers.tests.mock_brain as mb
 
 import numpy as np
-import os
 
 from mlagents.trainers.policy.tf_policy import TFPolicy
 from mlagents.trainers.components.bc.module import BCModule
@@ -10,6 +9,11 @@ from mlagents.trainers.settings import (
     TrainerSettings,
     BehavioralCloningSettings,
     NetworkSettings,
+)
+
+from mlagents.trainers.tests.dummy_config import (
+    DISCRETE_DEMO_PATH,
+    CONTINUOUS_DEMO_PATH,
 )
 
 
@@ -38,17 +42,13 @@ def create_bc_module(mock_behavior_specs, bc_settings, use_rnn, tanhresample):
 def test_bcmodule_defaults():
     # See if default values match
     mock_specs = mb.create_mock_3dball_behavior_specs()
-    bc_settings = BehavioralCloningSettings(
-        demo_path=os.path.dirname(os.path.abspath(__file__)) + "/" + "test.demo"
-    )
+    bc_settings = BehavioralCloningSettings(demo_path=CONTINUOUS_DEMO_PATH)
     bc_module = create_bc_module(mock_specs, bc_settings, False, False)
     assert bc_module.num_epoch == 3
     assert bc_module.batch_size == TrainerSettings().hyperparameters.batch_size
     # Assign strange values and see if it overrides properly
     bc_settings = BehavioralCloningSettings(
-        demo_path=os.path.dirname(os.path.abspath(__file__)) + "/" + "test.demo",
-        num_epoch=100,
-        batch_size=10000,
+        demo_path=CONTINUOUS_DEMO_PATH, num_epoch=100, batch_size=10000
     )
     bc_module = create_bc_module(mock_specs, bc_settings, False, False)
     assert bc_module.num_epoch == 100
@@ -59,9 +59,7 @@ def test_bcmodule_defaults():
 @pytest.mark.parametrize("is_sac", [True, False], ids=["sac", "ppo"])
 def test_bcmodule_update(is_sac):
     mock_specs = mb.create_mock_3dball_behavior_specs()
-    bc_settings = BehavioralCloningSettings(
-        demo_path=os.path.dirname(os.path.abspath(__file__)) + "/" + "test.demo"
-    )
+    bc_settings = BehavioralCloningSettings(demo_path=CONTINUOUS_DEMO_PATH)
     bc_module = create_bc_module(mock_specs, bc_settings, False, is_sac)
     stats = bc_module.update()
     for _, item in stats.items():
@@ -72,10 +70,7 @@ def test_bcmodule_update(is_sac):
 @pytest.mark.parametrize("is_sac", [True, False], ids=["sac", "ppo"])
 def test_bcmodule_constant_lr_update(is_sac):
     mock_specs = mb.create_mock_3dball_behavior_specs()
-    bc_settings = BehavioralCloningSettings(
-        demo_path=os.path.dirname(os.path.abspath(__file__)) + "/" + "test.demo",
-        steps=0,
-    )
+    bc_settings = BehavioralCloningSettings(demo_path=CONTINUOUS_DEMO_PATH, steps=0)
     bc_module = create_bc_module(mock_specs, bc_settings, False, is_sac)
     stats = bc_module.update()
     for _, item in stats.items():
@@ -90,9 +85,7 @@ def test_bcmodule_constant_lr_update(is_sac):
 @pytest.mark.parametrize("is_sac", [True, False], ids=["sac", "ppo"])
 def test_bcmodule_rnn_update(is_sac):
     mock_specs = mb.create_mock_3dball_behavior_specs()
-    bc_settings = BehavioralCloningSettings(
-        demo_path=os.path.dirname(os.path.abspath(__file__)) + "/" + "test.demo"
-    )
+    bc_settings = BehavioralCloningSettings(demo_path=CONTINUOUS_DEMO_PATH)
     bc_module = create_bc_module(mock_specs, bc_settings, True, is_sac)
     stats = bc_module.update()
     for _, item in stats.items():
@@ -103,9 +96,7 @@ def test_bcmodule_rnn_update(is_sac):
 @pytest.mark.parametrize("is_sac", [True, False], ids=["sac", "ppo"])
 def test_bcmodule_dc_visual_update(is_sac):
     mock_specs = mb.create_mock_banana_behavior_specs()
-    bc_settings = BehavioralCloningSettings(
-        demo_path=os.path.dirname(os.path.abspath(__file__)) + "/" + "testdcvis.demo"
-    )
+    bc_settings = BehavioralCloningSettings(demo_path=DISCRETE_DEMO_PATH)
     bc_module = create_bc_module(mock_specs, bc_settings, False, is_sac)
     stats = bc_module.update()
     for _, item in stats.items():
@@ -116,9 +107,7 @@ def test_bcmodule_dc_visual_update(is_sac):
 @pytest.mark.parametrize("is_sac", [True, False], ids=["sac", "ppo"])
 def test_bcmodule_rnn_dc_update(is_sac):
     mock_specs = mb.create_mock_banana_behavior_specs()
-    bc_settings = BehavioralCloningSettings(
-        demo_path=os.path.dirname(os.path.abspath(__file__)) + "/" + "testdcvis.demo"
-    )
+    bc_settings = BehavioralCloningSettings(demo_path=DISCRETE_DEMO_PATH)
     bc_module = create_bc_module(mock_specs, bc_settings, True, is_sac)
     stats = bc_module.update()
     for _, item in stats.items():
