+++ ml-agents/mlagents/trainers/tests/test_subprocess_env_manager.py
@@ -18,11 +18,11 @@ from mlagents_envs.exception import UnityEnvironmentException
 from mlagents.trainers.tests.simple_test_envs import SimpleEnvironment
 from mlagents.trainers.stats import StatsReporter
 from mlagents.trainers.agent_processor import AgentManagerQueue
-from mlagents.trainers.tests.test_simple_rl import (
-    _check_environment_trains,
-    PPO_CONFIG,
+from mlagents.trainers.tests.check_env_trains import (
+    check_environment_trains,
     DebugWriter,
 )
+from mlagents.trainers.tests.dummy_config import PPO_CONFIG
 
 
 def mock_env_factory(worker_id):
@@ -193,7 +193,7 @@ def test_subprocess_env_endtoend(num_envs):
         simple_env_factory, EngineConfig.default_config(), num_envs
     )
     # Run PPO using env_manager
-    _check_environment_trains(
+    check_environment_trains(
         simple_env_factory(0, []),
         {"1D": PPO_CONFIG},
         env_manager=env_manager,
