+++ ml-agents/mlagents/trainers/tests/torch/test_simple_rl.py
@@ -1,168 +1,44 @@
-import math
-import tempfile
 import pytest
-import numpy as np
 import attr
-from typing import Dict
+
 
 from mlagents.trainers.tests.simple_test_envs import (
     SimpleEnvironment,
     MemoryEnvironment,
     RecordEnvironment,
 )
-from mlagents.trainers.trainer_controller import TrainerController
-from mlagents.trainers.trainer_util import TrainerFactory
-from mlagents.trainers.simple_env_manager import SimpleEnvManager
+
 from mlagents.trainers.demo_loader import write_demo
-from mlagents.trainers.stats import StatsReporter, StatsWriter, StatsSummary
+
 from mlagents.trainers.settings import (
-    TrainerSettings,
-    PPOSettings,
-    SACSettings,
     NetworkSettings,
     SelfPlaySettings,
     BehavioralCloningSettings,
     GAILSettings,
-    TrainerType,
     RewardSignalType,
     EncoderType,
-    ScheduleType,
-    FrameworkType,
 )
-from mlagents.trainers.environment_parameter_manager import EnvironmentParameterManager
-from mlagents_envs.side_channel.environment_parameters_channel import (
-    EnvironmentParametersChannel,
-)
+
 from mlagents_envs.communicator_objects.demonstration_meta_pb2 import (
     DemonstrationMetaProto,
 )
 from mlagents_envs.communicator_objects.brain_parameters_pb2 import BrainParametersProto
 from mlagents_envs.communicator_objects.space_type_pb2 import discrete, continuous
 
+from mlagents.trainers.tests.dummy_config import PPO_CONFIG, SAC_CONFIG
+from mlagents.trainers.tests.check_env_trains import (
+    check_environment_trains,
+    default_reward_processor,
+)
+
 BRAIN_NAME = "1D"
-
-
-PPO_CONFIG = TrainerSettings(
-    trainer_type=TrainerType.PPO,
-    hyperparameters=PPOSettings(
-        learning_rate=5.0e-3,
-        learning_rate_schedule=ScheduleType.CONSTANT,
-        batch_size=16,
-        buffer_size=64,
-    ),
-    network_settings=NetworkSettings(num_layers=1, hidden_units=32),
-    summary_freq=500,
-    max_steps=3000,
-    threaded=False,
-    framework=FrameworkType.PYTORCH,
-)
-
-SAC_CONFIG = TrainerSettings(
-    trainer_type=TrainerType.SAC,
-    hyperparameters=SACSettings(
-        learning_rate=5.0e-3,
-        learning_rate_schedule=ScheduleType.CONSTANT,
-        batch_size=8,
-        buffer_init_steps=100,
-        buffer_size=5000,
-        tau=0.01,
-        init_entcoef=0.01,
-    ),
-    network_settings=NetworkSettings(num_layers=1, hidden_units=16),
-    summary_freq=100,
-    max_steps=1000,
-    threaded=False,
-    framework=FrameworkType.PYTORCH,
-)
-
-
-# The reward processor is passed as an argument to _check_environment_trains.
-# It is applied to the list of all final rewards for each brain individually.
-# This is so that we can process all final rewards in different ways for different algorithms.
-# Custom reward processors should be built within the test function and passed to _check_environment_trains
-# Default is average over the last 5 final rewards
-def default_reward_processor(rewards, last_n_rewards=5):
-    rewards_to_use = rewards[-last_n_rewards:]
-    # For debugging tests
-    print(f"Last {last_n_rewards} rewards:", rewards_to_use)
-    return np.array(rewards[-last_n_rewards:], dtype=np.float32).mean()
-
-
-class DebugWriter(StatsWriter):
-    """
-    Print to stdout so stats can be viewed in pytest
-    """
-
-    def __init__(self):
-        self._last_reward_summary: Dict[str, float] = {}
-
-    def get_last_rewards(self):
-        return self._last_reward_summary
-
-    def write_stats(
-        self, category: str, values: Dict[str, StatsSummary], step: int
-    ) -> None:
-        for val, stats_summary in values.items():
-            if val == "Environment/Cumulative Reward":
-                print(step, val, stats_summary.mean)
-                self._last_reward_summary[category] = stats_summary.mean
-
-
-def _check_environment_trains(
-    env,
-    trainer_config,
-    reward_processor=default_reward_processor,
-    env_parameter_manager=None,
-    success_threshold=0.9,
-    env_manager=None,
-):
-    if env_parameter_manager is None:
-        env_parameter_manager = EnvironmentParameterManager()
-    # Create controller and begin training.
-    with tempfile.TemporaryDirectory() as dir:
-        run_id = "id"
-        seed = 1337
-        StatsReporter.writers.clear()  # Clear StatsReporters so we don't write to file
-        debug_writer = DebugWriter()
-        StatsReporter.add_writer(debug_writer)
-        if env_manager is None:
-            env_manager = SimpleEnvManager(env, EnvironmentParametersChannel())
-        trainer_factory = TrainerFactory(
-            trainer_config=trainer_config,
-            output_path=dir,
-            train_model=True,
-            load_model=False,
-            seed=seed,
-            param_manager=env_parameter_manager,
-            multi_gpu=False,
-        )
-
-        tc = TrainerController(
-            trainer_factory=trainer_factory,
-            output_path=dir,
-            run_id=run_id,
-            param_manager=env_parameter_manager,
-            train=True,
-            training_seed=seed,
-        )
-
-        # Begin training
-        tc.start_learning(env_manager)
-        if (
-            success_threshold is not None
-        ):  # For tests where we are just checking setup and not reward
-            processed_rewards = [
-                reward_processor(rewards) for rewards in env.final_rewards.values()
-            ]
-            assert all(not math.isnan(reward) for reward in processed_rewards)
-            assert all(reward > success_threshold for reward in processed_rewards)
 
 
 @pytest.mark.parametrize("use_discrete", [True, False])
 def test_simple_ppo(use_discrete):
     env = SimpleEnvironment([BRAIN_NAME], use_discrete=use_discrete)
     config = attr.evolve(PPO_CONFIG)
-    _check_environment_trains(env, {BRAIN_NAME: config})
+    check_environment_trains(env, {BRAIN_NAME: config})
 
 
 @pytest.mark.parametrize("use_discrete", [True, False])
@@ -174,7 +50,7 @@ def test_2d_ppo(use_discrete):
         PPO_CONFIG.hyperparameters, batch_size=64, buffer_size=640
     )
     config = attr.evolve(PPO_CONFIG, hyperparameters=new_hyperparams, max_steps=10000)
-    _check_environment_trains(env, {BRAIN_NAME: config})
+    check_environment_trains(env, {BRAIN_NAME: config})
 
 
 @pytest.mark.parametrize("use_discrete", [True, False])
@@ -189,7 +65,7 @@ def test_visual_ppo(num_visual, use_discrete):
     )
     new_hyperparams = attr.evolve(PPO_CONFIG.hyperparameters, learning_rate=3.0e-4)
     config = attr.evolve(PPO_CONFIG, hyperparameters=new_hyperparams)
-    _check_environment_trains(env, {BRAIN_NAME: config})
+    check_environment_trains(env, {BRAIN_NAME: config})
 
 
 @pytest.mark.parametrize("num_visual", [1, 2])
@@ -215,7 +91,7 @@ def test_visual_advanced_ppo(vis_encode_type, num_visual):
         summary_freq=100,
     )
     # The number of steps is pretty small for these encoders
-    _check_environment_trains(env, {BRAIN_NAME: config}, success_threshold=0.5)
+    check_environment_trains(env, {BRAIN_NAME: config}, success_threshold=0.5)
 
 
 @pytest.mark.parametrize("use_discrete", [True, False])
@@ -234,14 +110,14 @@ def test_recurrent_ppo(use_discrete):
         network_settings=new_network_settings,
         max_steps=5000,
     )
-    _check_environment_trains(env, {BRAIN_NAME: config}, success_threshold=0.9)
+    check_environment_trains(env, {BRAIN_NAME: config}, success_threshold=0.9)
 
 
 @pytest.mark.parametrize("use_discrete", [True, False])
 def test_simple_sac(use_discrete):
     env = SimpleEnvironment([BRAIN_NAME], use_discrete=use_discrete)
     config = attr.evolve(SAC_CONFIG)
-    _check_environment_trains(env, {BRAIN_NAME: config})
+    check_environment_trains(env, {BRAIN_NAME: config})
 
 
 @pytest.mark.parametrize("use_discrete", [True, False])
@@ -251,7 +127,7 @@ def test_2d_sac(use_discrete):
     )
     new_hyperparams = attr.evolve(SAC_CONFIG.hyperparameters, buffer_init_steps=2000)
     config = attr.evolve(SAC_CONFIG, hyperparameters=new_hyperparams, max_steps=10000)
-    _check_environment_trains(env, {BRAIN_NAME: config}, success_threshold=0.8)
+    check_environment_trains(env, {BRAIN_NAME: config}, success_threshold=0.8)
 
 
 @pytest.mark.parametrize("use_discrete", [True, False])
@@ -268,7 +144,7 @@ def test_visual_sac(num_visual, use_discrete):
         SAC_CONFIG.hyperparameters, batch_size=16, learning_rate=3e-4
     )
     config = attr.evolve(SAC_CONFIG, hyperparameters=new_hyperparams)
-    _check_environment_trains(env, {BRAIN_NAME: config})
+    check_environment_trains(env, {BRAIN_NAME: config})
 
 
 @pytest.mark.parametrize("num_visual", [1, 2])
@@ -298,7 +174,7 @@ def test_visual_advanced_sac(vis_encode_type, num_visual):
         max_steps=100,
     )
     # The number of steps is pretty small for these encoders
-    _check_environment_trains(env, {BRAIN_NAME: config}, success_threshold=0.5)
+    check_environment_trains(env, {BRAIN_NAME: config}, success_threshold=0.5)
 
 
 @pytest.mark.parametrize("use_discrete", [True, False])
@@ -324,7 +200,7 @@ def test_recurrent_sac(use_discrete):
         network_settings=new_networksettings,
         max_steps=5000,
     )
-    _check_environment_trains(env, {BRAIN_NAME: config})
+    check_environment_trains(env, {BRAIN_NAME: config})
 
 
 @pytest.mark.parametrize("use_discrete", [True, False])
@@ -336,7 +212,7 @@ def test_simple_ghost(use_discrete):
         play_against_latest_model_ratio=1.0, save_steps=2000, swap_steps=2000
     )
     config = attr.evolve(PPO_CONFIG, self_play=self_play_settings, max_steps=2500)
-    _check_environment_trains(env, {BRAIN_NAME: config})
+    check_environment_trains(env, {BRAIN_NAME: config})
 
 
 @pytest.mark.parametrize("use_discrete", [True, False])
@@ -350,7 +226,7 @@ def test_simple_ghost_fails(use_discrete):
         play_against_latest_model_ratio=1.0, save_steps=2000, swap_steps=4000
     )
     config = attr.evolve(PPO_CONFIG, self_play=self_play_settings, max_steps=2500)
-    _check_environment_trains(env, {BRAIN_NAME: config}, success_threshold=None)
+    check_environment_trains(env, {BRAIN_NAME: config}, success_threshold=None)
     processed_rewards = [
         default_reward_processor(rewards) for rewards in env.final_rewards.values()
     ]
@@ -374,7 +250,7 @@ def test_simple_asymm_ghost(use_discrete):
         team_change=400,
     )
     config = attr.evolve(PPO_CONFIG, self_play=self_play_settings, max_steps=4000)
-    _check_environment_trains(env, {BRAIN_NAME: config, brain_name_opp: config})
+    check_environment_trains(env, {BRAIN_NAME: config, brain_name_opp: config})
 
 
 @pytest.mark.parametrize("use_discrete", [True, False])
@@ -393,7 +269,7 @@ def test_simple_asymm_ghost_fails(use_discrete):
         team_change=2000,
     )
     config = attr.evolve(PPO_CONFIG, self_play=self_play_settings, max_steps=3000)
-    _check_environment_trains(
+    check_environment_trains(
         env, {BRAIN_NAME: config, brain_name_opp: config}, success_threshold=None
     )
     processed_rewards = [
@@ -451,7 +327,7 @@ def test_gail(simple_record, use_discrete, trainer_config):
         behavioral_cloning=bc_settings,
         max_steps=500,
     )
-    _check_environment_trains(env, {BRAIN_NAME: config}, success_threshold=0.9)
+    check_environment_trains(env, {BRAIN_NAME: config}, success_threshold=0.9)
 
 
 @pytest.mark.parametrize("use_discrete", [True, False])
@@ -476,7 +352,7 @@ def test_gail_visual_ppo(simple_record, use_discrete):
         behavioral_cloning=bc_settings,
         max_steps=1000,
     )
-    _check_environment_trains(env, {BRAIN_NAME: config}, success_threshold=0.9)
+    check_environment_trains(env, {BRAIN_NAME: config}, success_threshold=0.9)
 
 
 @pytest.mark.parametrize("use_discrete", [True, False])
@@ -503,4 +379,4 @@ def test_gail_visual_sac(simple_record, use_discrete):
         behavioral_cloning=bc_settings,
         max_steps=500,
     )
-    _check_environment_trains(env, {BRAIN_NAME: config}, success_threshold=0.9)
+    check_environment_trains(env, {BRAIN_NAME: config}, success_threshold=0.9)
